requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
betterproto==2.0.0b7
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
dacite==1.9.2
ffmpy==0.6.3
Flask==3.1.2
frozenlist==1.8.0
grpclib==0.4.8
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mashumaro==3.17
multidict==6.7.0
obs-websocket-py==1.0
obsws-python==1.8.0
propcache==0.4.1
protobuf==6.32.1
protobuf3-to-dict==0.1.5
pyee==13.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-socks==2.7.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
TikTokLive==6.6.5
typing_extensions==4.15.0
urllib3==2.5.0
websocket-client==1.9.0
websockets==15.0.1
websockets_proxy==0.1.3
Werkzeug==3.1.3
yarl==1.22.0